+++ modules/AWSPowerShell/Cmdlets/CostExplorer/Basic/New-CECostCategoryDefinition-Cmdlet.cs
@@ -66,6 +66,22 @@ namespace Amazon.PowerShell.Cmdlets.CE
         public System.String Name { get; set; }
         #endregion
         
+        #region Parameter ResourceTag
+        /// <summary>
+        /// <para>
+        /// <para> An optional list of tags to associate with the specified <a href="https://docs.aws.amazon.com/aws-cost-management/latest/APIReference/API_CostCategory.html"><code>CostCategory</code></a>. You can use resource tags to control access to your
+        /// <code>cost category</code> using IAM policies.</para><para>Each tag consists of a key and a value, and each key must be unique for the resource.
+        /// The following restrictions apply to resource tags:</para><ul><li><para>Although the maximum number of array members is 200, you can assign a maximum of 50
+        /// user-tags to one resource. The remaining are reserved for Amazon Web Services use</para></li><li><para>The maximum length of a key is 128 characters</para></li><li><para>The maximum length of a value is 256 characters</para></li><li><para>Valid characters for keys and values are: <code>A-Z</code>, <code>a-z</code>, spaces,
+        /// <code>_.:/=+-</code></para></li><li><para>Keys and values are case sensitive</para></li><li><para>Keys and values are trimmed for any leading or trailing whitespaces</para></li><li><para>Don’t use <code>aws:</code> as a prefix for your keys. This prefix is reserved for
+        /// Amazon Web Services use</para></li></ul>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        [Alias("ResourceTags")]
+        public Amazon.CostExplorer.Model.ResourceTag[] ResourceTag { get; set; }
+        #endregion
+        
         #region Parameter Rule
         /// <summary>
         /// <para>
@@ -182,6 +198,10 @@ namespace Amazon.PowerShell.Cmdlets.CE
                 WriteWarning("You are passing $null as a value for parameter Name which is marked as required. In case you believe this parameter was incorrectly marked as required, report this by opening an issue at https://github.com/aws/aws-tools-for-powershell/issues.");
             }
             #endif
+            if (this.ResourceTag != null)
+            {
+                context.ResourceTag = new List<Amazon.CostExplorer.Model.ResourceTag>(this.ResourceTag);
+            }
             if (this.Rule != null)
             {
                 context.Rule = new List<Amazon.CostExplorer.Model.CostCategoryRule>(this.Rule);
@@ -226,6 +246,10 @@ namespace Amazon.PowerShell.Cmdlets.CE
             if (cmdletContext.Name != null)
             {
                 request.Name = cmdletContext.Name;
+            }
+            if (cmdletContext.ResourceTag != null)
+            {
+                request.ResourceTags = cmdletContext.ResourceTag;
             }
             if (cmdletContext.Rule != null)
             {
@@ -302,6 +326,7 @@ namespace Amazon.PowerShell.Cmdlets.CE
         {
             public System.String DefaultValue { get; set; }
             public System.String Name { get; set; }
+            public List<Amazon.CostExplorer.Model.ResourceTag> ResourceTag { get; set; }
             public List<Amazon.CostExplorer.Model.CostCategoryRule> Rule { get; set; }
             public Amazon.CostExplorer.CostCategoryRuleVersion RuleVersion { get; set; }
             public List<Amazon.CostExplorer.Model.CostCategorySplitChargeRule> SplitChargeRule { get; set; }