+++ modules/AWSPowerShell/Cmdlets/ECS/Basic/Update-ECSService-Cmdlet.cs
@@ -280,7 +280,18 @@ namespace Amazon.PowerShell.Cmdlets.ECS
         /// name, the container name, and the container port to access from the load balancer.
         /// The container name is as it appears in a container definition.</para><para>When you add, update, or remove a load balancer configuration, Amazon ECS starts new
         /// tasks with the updated Elastic Load Balancing configuration, and then stops the old
-        /// tasks when the new tasks are running.</para><para>You can remove existing <code>loadBalancers</code> by passing an empty list.</para>
+        /// tasks when the new tasks are running.</para><para>For services that use rolling updates, you can add, update, or remove Elastic Load
+        /// Balancing target groups. You can update from a single target group to multiple target
+        /// groups and from multiple target groups to a single target group.</para><para>For services that use blue/green deployments, you can update Elastic Load Balancing
+        /// target groups by using <code><a href="https://docs.aws.amazon.com/codedeploy/latest/APIReference/API_CreateDeployment.html">CreateDeployment</a></code> through CodeDeploy. Note that multiple target groups are not supported for
+        /// blue/green deployments. For more information see <a href="https://docs.aws.amazon.com/AmazonECS/latest/developerguide/register-multiple-targetgroups.html">Register
+        /// multiple target groups with a service</a> in the <i>Amazon Elastic Container Service
+        /// Developer Guide</i>. </para><para>For services that use the external deployment controller, you can add, update, or
+        /// remove load balancers by using <a href="https://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_CreateTaskSet.html">CreateTaskSet</a>.
+        /// Note that multiple target groups are not supported for external deployments. For more
+        /// information see <a href="https://docs.aws.amazon.com/AmazonECS/latest/developerguide/register-multiple-targetgroups.html">Register
+        /// multiple target groups with a service</a> in the <i>Amazon Elastic Container Service
+        /// Developer Guide</i>. </para><para>You can remove existing <code>loadBalancers</code> by passing an empty list.</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]